requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
python-dotenv==1.0.0
pytest==7.4.3
logging==0.4.9.6
